Decompile in background thread

+++ ulyp-ui-repository/src/main/java/com/ulyp/core/impl/FileBasedCallRecordDatabase.java
@@ -259,7 +259,7 @@ public class FileBasedCallRecordDatabase implements CallRecordDatabase {
 
     @Override
     public long getSubtreeCount(long id) {
-        return idToSubtreeCountMap.get(id);
+        return idToSubtreeCountMap.get((int) id);
     }
 
     @Override
